More tweaks

+++ AutoModTests/LogicTests.cs
@@ -10,9 +10,9 @@ public static class LogicTests
 {
     [Theory]
     [InlineData(Pikachu, 1)]
-    public static void TestShinyLockInfo(Species species, int form, bool locked = true)
+    public static void TestShinyLockInfo(Species species, byte form, bool locked = true)
     {
-        var result = SimpleEdits.IsShinyLockedSpeciesForm((int)species, form);
+        var result = SimpleEdits.IsShinyLockedSpeciesForm((ushort)species, form);
         result.Should().Be(locked);
     }
 }
